Move event definitions to new module
